backend/requirements.txt
Flask==2.0.1
kafka-python==2.0.2
numpy==1.23.5
opencv-python==4.8.1.78
Werkzeug==2.0.2
scikit-learn==1.2.2
pandas==1.5.3
requests==2.28.2
Pillow==10.1.0
